requirements.txt
Flask==2.3.3
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.1.78
matplotlib==3.7.2
scikit-learn==1.3.0                             